finish flow described in task

+++ search/src/main/java/pl/wojtek/ask/data/Repository.kt
@@ -6,4 +6,4 @@ package pl.wojtek.ask.data
 
 
 data class Repository(val repositoryName:String,val urlToRepo:String,
-                      val score:Double, val ownerName:String, val id:Int=0 )
+                      val score:Int, val ownerName:String, val id:Int=0 )
